libft

+++ libft/ft_bzero.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-void	ft_bzero(void *ptr, unsigned int n)
+void	ft_bzero(void *ptr, size_t n)
 {
 	unsigned char	*p;
