+++ libft/ft_calloc.c
@@ -11,9 +11,9 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-void	*ft_calloc(unsigned int n, unsigned int size)
+void	*ft_calloc(size_t n, size_t size)
 {
-	unsigned long	total;
+	size_t			total;
 	unsigned int	i;
 	void			*alloc;
