+++ libft/ft_memcmp.c
@@ -11,7 +11,7 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-int	ft_memcmp(const void *r1, const void *r2, unsigned int n)
+int	ft_memcmp(const void *r1, const void *r2, size_t n)
 {
 	unsigned char	*_r1;
 	unsigned char	*_r2;
