+++ libft/ft_strlen.c
@@ -11,9 +11,9 @@
 /* ************************************************************************** */
 #include "libft.h"
 
-unsigned int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	unsigned int	l;
+	size_t	l;
 
 	l = 0;
 	while (str[l])
