+++ libft/libft.h
@@ -13,6 +13,7 @@
 # define LIBFT_H
 # include <stdlib.h>
 # include <unistd.h>
+# include <sys/types.h>
 
 int				ft_isalpha(int c);
 int				ft_isdigit(int c);
@@ -21,24 +22,24 @@ int				ft_isascii(int c);
 int				ft_isprint(int c);
 int				ft_tolower(int c);
 int				ft_toupper(int c);
-unsigned int	ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 char			*ft_strdup(const char *str);
 char			*ft_strchr(const char *haystack, int needle);
 char			*ft_strrchr(const char *haystack, int needle);
-unsigned int	ft_strlcpy(char *dest, const char *src, unsigned int n);
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int n);
-int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
+size_t			ft_strlcpy(char *dest, const char *src, size_t n);
+size_t			ft_strlcat(char *dest, const char *src, size_t n);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
 char			*ft_strnstr(const char *haystack, const char *needle,
-					unsigned int n);
+					size_t n);
 char			*ft_strtrim(const char *str, const char *set);
 char			**ft_split(const char *str, char set);
-void			ft_bzero(void *ptr, unsigned int n);
-void			*ft_memset(void *s, int c, unsigned int n);
-void			*ft_memcpy(void *dest, const void *src, unsigned int n);
-void			*ft_memmove(void *dest, const void *src, unsigned int n);
-void			*ft_memchr(const void *ptr, int c, unsigned int n);
-int				ft_memcmp(const void *r1, const void *r2, unsigned int n);
-void			*ft_calloc(unsigned int n, unsigned int size);
+void			ft_bzero(void *ptr, size_t n);
+void			*ft_memset(void *s, int c, size_t n);
+void			*ft_memcpy(void *dest, const void *src, size_t n);
+void			*ft_memmove(void *dest, const void *src, size_t n);
+void			*ft_memchr(const void *ptr, int c, size_t n);
+int				ft_memcmp(const void *r1, const void *r2, size_t n);
+void			*ft_calloc(size_t n, size_t size);
 int				ft_atoi(const char *str);
 void			ft_putchar_fd(char c, int fd);
 void			ft_putstr_fd(char *str, int fd);
